+++ src/apiskillqueue.cc
@@ -148,8 +148,8 @@ ApiSkillQueue::get_spph_for_current (void) const
   if (item == 0)
     return 0;
 
-  double skill_sp = item->end_sp - item->start_sp;
-  double train_time = item->end_time_t - item->start_time_t;
+  double skill_sp = (double)(item->end_sp - item->start_sp);
+  double train_time = (double)(item->end_time_t - item->start_time_t);
   return (unsigned int)(skill_sp * 3600.0 / train_time);
 }
